+++ android/src/main/java/com/roughike/facebooklogin/facebooklogin/FacebookLoginPlugin.java
@@ -1,5 +1,9 @@
 package com.roughike.facebooklogin.facebooklogin;
 
+import android.app.Activity;
+
+import androidx.annotation.NonNull;
+
 import com.facebook.AccessToken;
 import com.facebook.CallbackManager;
 import com.facebook.login.LoginBehavior;
@@ -8,14 +12,17 @@ import com.facebook.login.LoginManager;
 import java.util.List;
 import java.util.Map;
 
+import io.flutter.embedding.engine.plugins.FlutterPlugin;
+import io.flutter.embedding.engine.plugins.activity.ActivityAware;
+import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
 import io.flutter.plugin.common.MethodChannel.Result;
 import io.flutter.plugin.common.PluginRegistry.Registrar;
 
-public class FacebookLoginPlugin implements MethodCallHandler {
-    private static final String CHANNEL_NAME = "com.roughike/flutter_facebook_login";
+public class FacebookLoginPlugin implements FlutterPlugin, MethodCallHandler {
+    private static final String CHANNEL_NAME = "com.roughike/meta_facebook_login";
 
     private static final String ERROR_UNKNOWN_LOGIN_BEHAVIOR = "unknown_login_behavior";
 
@@ -32,15 +39,10 @@ public class FacebookLoginPlugin implements MethodCallHandler {
     private static final String LOGIN_BEHAVIOR_WEB_VIEW_ONLY = "webViewOnly";
 
     private final FacebookSignInDelegate delegate;
+    private MethodChannel channel;
 
-    private FacebookLoginPlugin(Registrar registrar) {
-        delegate = new FacebookSignInDelegate(registrar);
-    }
-
-    public static void registerWith(Registrar registrar) {
-        final FacebookLoginPlugin plugin = new FacebookLoginPlugin(registrar);
-        final MethodChannel channel = new MethodChannel(registrar.messenger(), CHANNEL_NAME);
-        channel.setMethodCallHandler(plugin);
+    private FacebookLoginPlugin() {
+        delegate = new FacebookSignInDelegate();
     }
 
     @Override
@@ -89,20 +91,32 @@ public class FacebookLoginPlugin implements MethodCallHandler {
         }
     }
 
-    public static final class FacebookSignInDelegate {
-        private final Registrar registrar;
+    @Override
+    public void onAttachedToEngine(@NonNull FlutterPluginBinding binding) {
+        final FacebookLoginPlugin plugin = new FacebookLoginPlugin();
+        channel = new MethodChannel(binding.getBinaryMessenger(), CHANNEL_NAME);
+
+        channel.setMethodCallHandler(plugin);
+    }
+
+    @Override
+    public void onDetachedFromEngine(@NonNull FlutterPluginBinding binding) {
+        channel.setMethodCallHandler(null);
+    }
+
+    public static final class FacebookSignInDelegate implements ActivityAware {
         private final CallbackManager callbackManager;
         private final LoginManager loginManager;
         private final FacebookLoginResultDelegate resultDelegate;
 
-        public FacebookSignInDelegate(Registrar registrar) {
-            this.registrar = registrar;
+        private Activity activity;
+
+        public FacebookSignInDelegate() {
             this.callbackManager = CallbackManager.Factory.create();
             this.loginManager = LoginManager.getInstance();
             this.resultDelegate = new FacebookLoginResultDelegate(callbackManager);
 
             loginManager.registerCallback(callbackManager, resultDelegate);
-            registrar.addActivityResultListener(resultDelegate);
         }
 
         public void logIn(
@@ -110,7 +124,7 @@ public class FacebookLoginPlugin implements MethodCallHandler {
             resultDelegate.setPendingResult(METHOD_LOG_IN, result);
 
             loginManager.setLoginBehavior(loginBehavior);
-            loginManager.logIn(registrar.activity(), permissions);
+            if(activity != null) loginManager.logIn(activity, permissions);
         }
 
         public void logOut(Result result) {
@@ -123,6 +137,27 @@ public class FacebookLoginPlugin implements MethodCallHandler {
             Map<String, Object> tokenMap = FacebookLoginResults.accessToken(accessToken);
 
             result.success(tokenMap);
+        }
+
+        @Override
+        public void onAttachedToActivity(@NonNull ActivityPluginBinding binding) {
+            activity = binding.getActivity();
+            binding.addActivityResultListener(resultDelegate);
+        }
+
+        @Override
+        public void onDetachedFromActivityForConfigChanges() {
+            activity = null;
+        }
+
+        @Override
+        public void onReattachedToActivityForConfigChanges(@NonNull ActivityPluginBinding binding) {
+            activity = binding.getActivity();
+        }
+
+        @Override
+        public void onDetachedFromActivity() {
+            activity = null;
         }
     }
 }